+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * fix(ContextMenu/DropdownMenu): allow item content class override
 * fix(CommandPalette/ContextMenu/DropdownMenu): ensure items truncate work & itemTrailingIcon color
 * fix(ContentSearch): de-duplicate description and suffix
+* feat(ProseImg): improve `zoom` transition
 
 ## [2.0.0](https://github.com/bitrix24/b24ui/compare/v1.0.4...v2.0.0) (2025-10-21)